python-common, orch: Move ServiceSpec (+deps) to python-common
